feat: add test case list page and reactgrid dependencies
implemented a basic functionality to fetch test-cases from the
server, along with the hooks required. this changeset also includes
addition of reactgrid which will be used to list/display test-cases
when testers need to interact with them.

Some documentation updates on the structure of the UI navigation,
modeled as plantul mindmap